+++ tests/tests_pytorch/utilities/test_seed.py
@@ -1,6 +1,8 @@
 import os
+from typing import Mapping
 import random
 from unittest import mock
+from unittest.mock import MagicMock
 
 import numpy as np
 import pytest
@@ -96,3 +98,20 @@ def test_isolate_rng():
     with isolate_rng():
         generated = [random.random() for _ in range(3)]
     assert random.random() == generated[0]
+
+
+@mock.patch("pytorch_lightning.utilities.seed.log.info")
+@pytest.mark.parametrize("env_vars", [{"RANK": "0"}, {"RANK": "1"}, {"RANK": "4"}])
+@pytest.mark.parametrize("message", ["log message"])
+def test_seed_everything_log_info(log_mock: MagicMock, env_vars: Mapping[str, str], message: str):
+    """Test that if rank more then zero, add a prefix."""
+    with mock.patch.dict(os.environ, env_vars):
+        from pytorch_lightning.utilities.rank_zero import _get_rank
+
+        rank = _get_rank()
+
+        seed_utils._log_info(message=message)
+
+    expected_log = f"[rank: {rank}] {message}" if rank > 0 else message
+
+    log_mock.assert_called_once_with(expected_log)